Remove Theia extension related code
