--- components/gitpod-protocol/src/theia-plugins.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) 2020 Gitpod GmbH. All rights reserved.
- * Licensed under the GNU Affero General Public License (AGPL).
- * See License-AGPL.txt in the project root for license information.
- */
-
-export interface PluginIndexEntry {
-    loc: string
-    name: string
-    url: string
-}
