+++ components/server/src/workspace/config-provider.ts
@@ -17,7 +17,6 @@ import { MaybeContent } from "../repohost/file-provider";
 import { ConfigurationService } from "../config/configuration-service";
 import { HostContextProvider } from "../auth/host-context-provider";
 import { AuthorizationService } from "../user/authorization-service";
-import { TheiaPluginService } from "../theia-plugin/theia-plugin-service";
 import { TraceContext } from "@gitpod/gitpod-protocol/lib/util/tracing";
 import { Config } from "../config";
 
@@ -35,7 +34,6 @@ export class ConfigProvider {
     @inject(GitpodFileParser) protected readonly gitpodParser: GitpodFileParser;
     @inject(HostContextProvider) protected readonly hostContextProvider: HostContextProvider;
     @inject(AuthorizationService) protected readonly authService: AuthorizationService;
-    @inject(TheiaPluginService) protected readonly pluginService: TheiaPluginService;
     @inject(ProjectDB) protected readonly projectDB: ProjectDB;
     @inject(Config) protected readonly config: Config;
     @inject(ConfigurationService) protected readonly configurationService: ConfigurationService;
